♻️ Refactor: 훅 단순화 및 타입 재사용 정리
